+++ Sources/Core/Components/PlayerPlugin.swift
@@ -50,6 +50,14 @@ public protocol PlayerPlugin {
     ///
     func didLoad(media: PlayerMedia, duration: Double?)
     
+    ///
+    /// Called when the media has changed
+    /// - parameters:
+    ///    - media: PlayerMedia just loaded
+    ///    - previousMedia: PlayerMedia loaded before
+    ///
+    func didMediaChanged(_ media: PlayerMedia, previousMedia: PlayerMedia?)
+    
     /// Called when the media is playing
     func didStartPlaying()