+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 ---
-## [0.11.0]
+## [X.X.X]
 * Feature:
 	* Add item to play end time information to delegate and plugin
-	* Add did change media information to plugin
+	* Add did media changed information to plugin
 
 ## [0.10.0]
 * Breaking Change: