+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [X.X.X]
 * Feature:
 	* Add item to play end time information to delegate and plugin
+	* Add did media changed information to plugin
 
 ## [0.10.0]
 * Breaking Change: